requirements.txt
requests>=2.25.1
python-dotenv>=0.21.0
datetime
